+++ gherkin/src/main/java/io/cucumber/core/gherkin/FeatureParser.java
@@ -1,12 +1,13 @@
 package io.cucumber.core.gherkin;
 
 import java.net.URI;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.function.Supplier;
 
 public interface FeatureParser {
 
-    Feature parse(URI path, String source, Supplier<UUID> idGenerator);
+    Optional<Feature> parse(URI path, String source, Supplier<UUID> idGenerator);
 
     String version();